requirements.txt
google-generativeai
python-dotenv
sounddevice
numpy
librosa
soundfile
websockets
pynput